feat: admin user, error handler, custom page request

+++ ewm/src/main/java/ru/practicum/ewm/user/service/UserService.java
@@ -11,5 +11,5 @@ public interface UserService {
 
     void deleteUser(long id);
 
-    List<GetUserDto> getUsers(List<Long> ids, long from, long size);
+    List<GetUserDto> getUsers(List<Long> ids, int from, int size);
 }
